requirements.txt
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.2
mysqlclient==2.2.4
redis=5.0.7
